feat: v0.9.3 - M5 Pro 48GB RAM optimization, 23 agentic MCP tools, and 88.2% test coverage

+++ quanta/core/equivalence.py
@@ -119,4 +119,4 @@ def fidelity(
     dim = u_a.shape[0]
     trace = np.abs(np.trace(u_a.conj().T @ u_b))
 
-    return trace / dim
+    return float(trace / dim)
